fix: remove circular dependencies
Move `fetchNegativeUNL`, `fetchQuorum`, and `fetchMetrics` out of
`utils.js` and into `Streams.jsx` (the only place they were used).

This removes 387 circular dependencies.
